+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## Version `1.0.2`
 
+- `worxpace` Upgrade `@clerk/nextjs` to `5.2.12` & fix auth
 - `storybook` Add `Select (custom)`
 - `ui` Add `Select (custom)`, `Badge`